update script submodule structures
